Add JS router for lexicon page
Adds a new JS router to manage the Lexicon React app. This router ties
into the `Lexicon.API` module, so that a given route match will
automatically trigger the corresponding API call, enabling directly
navigating to URLs denoting a particular application state.

Adds a new `Link` component to update history whenever a link is
clicked.

Updates the Rails router to check for `Accept` headers to determine
whether to send JSON or simply send the user (with the full stateful
URL) to the app initialization page.