Lexicon: Add better handling of empty fields
Adds better handling of empty (valueless) fields in the lexicon entry
edit view. Now the lexicon will properly set the state of a given
formatter to `null` if there is no data, and will clear out the display
(showing a default view) if there is no data available. Clicking on the
placeholder will automatically create the minimum data structure needed
to begin editing.