Refactor lexicons to use cleaner formatter structure
Refactors the lexicon code to use a cleaner, more consistent formatter
interface that is more flexible and no longer treats associations as
exceptional entities that must be dealt with in a separate step from
columns in the lexicon entry itself.

Isolates formatters into individual files in the formatter namespace.

Removes need for unnecessary crossreference join models, as well as
unnecessary `updated_at` column in cross reference tables (since entries
are always either created or deleted).

Fixes a bug causing the URL not to update after created a new lexicon
entry.