Add lexicon search support for part of speech
Restores support for searching by part of speech given the new lexicon
schema structure. This entails the creation of a new `base_dataset`
concept used for all searches that adds in whatever computed values are
needed for the string comparison to succeed. Since the old setup
expected a column called `part_of_speech`, all that is needed is a
computed column of that name with the concatenated strings that form the
associated part of speech.